Work on sending packets

+++ include/Types.h
@@ -22,7 +22,7 @@ namespace Game3 {
 	using HitPoints    = uint32_t;
 	/** 1-based. */
 	using Layer        = uint8_t;
-	using PacketID     = uint32_t;
+	using PacketID     = uint16_t;
 	using Version      = uint32_t;
 	using GlobalID     = uint64_t;
